movie_organizer/requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
av==14.3.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cinemagoer==2023.5.1
click==8.1.8
color-matcher==0.6.0
comfyui_frontend_package==1.16.9
comfyui_workflow_templates==0.1.3
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
DateTime==5.5
ddt==1.7.2
Deprecated==1.2.18
diffusers==0.33.1
docutils==0.21.2
einops==0.8.1
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
IMDbPY==2022.7.9
importlib_metadata==8.6.1
Jinja2==3.1.6
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lark==1.2.2
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matrix-client==0.4.0
mdurl==0.1.2
mpmath==1.3.0
mss==10.0.0
multidict==6.4.3
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
packaging==25.0
piexif==1.1.3
pillow==11.2.1
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pycparser==2.22
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rawpy==0.25.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scipy==1.15.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
spandrel==0.4.1
SQLAlchemy==2.0.43
sympy==1.13.1
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchsde==0.2.6
torchvision==0.21.0
tqdm==4.67.1
trampoline==0.1.2
transformers==4.51.3
typer==0.15.2
typing_extensions==4.13.2
urllib3==1.26.20
uv==0.6.14
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
zope.interface==7.2
